requirements.txt
beautifulsoup4==4.6.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
idna==2.10
JPype1==1.0.2
konlpy==0.5.2
lxml==4.5.2
numpy==1.19.2
oauthlib==3.1.0
PySocks==1.7.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tqdm-conan==4.27.0
tweepy==3.9.0
urllib3==1.25.10
